WIP -> message box refactor
